Switch to api.Relation.create to support Oracle

+++ integration_tests/models/fact_order_line.sql
@@ -11,7 +11,7 @@
 
 SELECT
   lineitem.*,
-  cast(TO_CHAR(o_orderdate, 'YYYYMMDD') AS NUMBER(10,0)) AS o_orderdate_key,
+  cast(TO_CHAR(o_orderdate, 'YYYYMMDD') AS INTEGER) AS o_orderdate_key,
   coalesce(cast(l_orderkey as varchar(100)), '') || '~' || coalesce(cast(l_linenumber as varchar(100)), '') AS integration_id
 FROM {{ source('tpc_h', 'lineitem') }} lineitem
 JOIN {{ source('tpc_h', 'orders') }} orders ON l_orderkey = o_orderkey
